# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Class:
    properties:
      capacity:
        type: integer
      course:
        $ref: '#/definitions/models.Course'
      course_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        description: 'Contoh: "Basic English A"'
        type: string
      schedule:
        description: 'Contoh: "Senin & Rabu, 19:00 - 21:00"'
        type: string
      status:
        description: '''open'', ''full'', ''closed'''
        type: string
      updated_at:
        type: string
    type: object
  models.ClassPlacement:
    properties:
      class:
        $ref: '#/definitions/models.Class'
      class_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      placement_date:
        type: string
      registration:
        $ref: '#/definitions/models.Registration'
      registration_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Course:
    properties:
      classes:
        items:
          $ref: '#/definitions/models.Class'
        type: array
      created_at:
        type: string
      description:
        type: string
      duration:
        description: 'Contoh: "3 Bulan", "48 Jam"'
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      status:
        description: '''active'', ''inactive'''
        type: string
      updated_at:
        type: string
    type: object
  models.Payment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      payment_date:
        type: string
      payment_method:
        description: '''bank_transfer'', ''ewallet'', ''credit_card'''
        type: string
      registration:
        $ref: '#/definitions/models.Registration'
      registration_id:
        type: integer
      status:
        description: '''pending'', ''paid'', ''failed'', ''refunded'''
        type: string
      updated_at:
        type: string
    type: object
  models.Registration:
    properties:
      class_placement:
        $ref: '#/definitions/models.ClassPlacement'
      course:
        $ref: '#/definitions/models.Course'
      course_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      payment:
        $ref: '#/definitions/models.Payment'
      registration_date:
        type: string
      status:
        description: '''pending'', ''registered'', ''cancelled'', ''completed'''
        type: string
      student:
        $ref: '#/definitions/models.Student'
      student_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Student:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  services.CreateRegistrationRequest:
    properties:
      course_id:
        type: integer
      student_id:
        type: integer
    required:
    - course_id
    - student_id
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@englishcourse.com
    name: Novendra
  description: RESTful API backend untuk pendaftaran kursus bahasa Inggris dengan
    clean layered architecture.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: English Course Registration API
  version: "1.0"
paths:
...
  /courses/{id}/registrations:
    get:
      description: Mengembalikan semua siswa yang terdaftar di course tertentu
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Registration'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil daftar pendaftaran pada course
      tags:
      - Courses
...
  /registrations:
    get:
      description: Mengembalikan semua data registrasi beserta relasi student, course,
        dan payment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Registration'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil daftar seluruh pendaftaran
      tags:
      - Registrations
    post:
      consumes:
      - application/json
      description: Mendaftarkan siswa ke kursus dan secara otomatis membuat tagihan
        Payment pending
      parameters:
      - description: Payload pendaftaran
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Registration'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mendaftar ke kursus (Registration)
      tags:
      - Registrations
  /registrations/{id}:
    get:
      description: Mengembalikan detail registrasi berdasarkan ID
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Registration'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil detail pendaftaran
      tags:
      - Registrations
  /registrations/{id}/cancel:
    put:
      description: Mengubah status registrasi menjadi cancelled jika belum selesai
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Membatalkan pendaftaran
      tags:
      - Registrations
...
  /students/{id}/registrations:
    get:
      description: Mengembalikan daftar kursus yang didaftarkan oleh student tertentu
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Registration'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil riwayat pendaftaran student
      tags:
      - Students
swagger: "2.0"